Pull most of dealing with fig into separate file
