+++ extensions/terminal-suggest/src/terminalSuggestMain.ts
@@ -2,29 +2,26 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import * as vscode from 'vscode';
+
+import { ExecOptionsWithStringEncoding } from 'child_process';
 import * as fs from 'fs/promises';
 import * as path from 'path';
-import { ExecOptionsWithStringEncoding } from 'child_process';
-import { upstreamSpecs } from './constants';
-import codeCompletionSpec from './completions/code';
+import * as vscode from 'vscode';
 import cdSpec from './completions/cd';
+import codeCompletionSpec from './completions/code';
 import codeInsidersCompletionSpec from './completions/code-insiders';
+import { upstreamSpecs } from './constants';
+import { PathExecutableCache } from './env/pathExecutableCache';
 import { osIsWindows } from './helpers/os';
-import type { ICompletionResource } from './types';
+import { getFriendlyResourcePath } from './helpers/uri';
 import { getBashGlobals } from './shell/bash';
-import { getZshGlobals } from './shell/zsh';
 import { getFishGlobals } from './shell/fish';
 import { getPwshGlobals } from './shell/pwsh';
+import { getZshGlobals } from './shell/zsh';
 import { getTokenType, TokenType } from './tokens';
-import { PathExecutableCache } from './env/pathExecutableCache';
-import { getFriendlyResourcePath } from './helpers/uri';
-import { ArgumentParserResult, parseArguments } from './fig/autocomplete-parser/parseArguments';
-import { getCommand, type Command } from './fig/shell-parser/command';
-import { SuggestionFlag } from './fig/shared/utils';
-import { createGeneratorState } from './fig/autocomplete/state/generators';
-import { Visibility, type AutocompleteState } from './fig/autocomplete/state/types';
-import type { FigState } from './fig/autocomplete/fig/hooks';
+import type { ICompletionResource } from './types';
+import { createCompletionItem } from './helpers/completionItem';
+import { getFigSuggestionLabel, getFigSuggestions, getFixSuggestionDescription } from './fig/figInterface';
 
 // TODO: remove once API is finalized
 export const enum TerminalShellType {
@@ -171,38 +168,6 @@ export async function resolveCwdFromPrefix(prefix: string, currentCwd?: vscode.U
 	// If the prefix is not a folder, return the current cwd
 	return currentCwd;
 }
-function getDescription(spec: Fig.Spec): string {
-	if ('description' in spec) {
-		return spec.description ?? '';
-	}
-	return '';
-}
-
-function getLabel(spec: Fig.Spec | Fig.Arg | Fig.Suggestion | string): string[] | undefined {
-	if (typeof spec === 'string') {
-		return [spec];
-	}
-	if (typeof spec.name === 'string') {
-		return [spec.name];
-	}
-	if (!Array.isArray(spec.name) || spec.name.length === 0) {
-		return;
-	}
-	return spec.name;
-}
-
-function createCompletionItem(cursorPosition: number, prefix: string, commandResource: ICompletionResource, detail?: string, documentation?: string | vscode.MarkdownString, kind?: vscode.TerminalCompletionItemKind): vscode.TerminalCompletionItem {
-	const endsWithSpace = prefix.endsWith(' ');
-	const lastWord = endsWithSpace ? '' : prefix.split(' ').at(-1) ?? '';
-	return {
-		label: commandResource.label,
-		detail: detail ?? commandResource.detail ?? '',
-		documentation,
-		replacementIndex: cursorPosition - lastWord.length,
-		replacementLength: lastWord.length,
-		kind: kind ?? commandResource.kind ?? vscode.TerminalCompletionItemKind.Method
-	};
-}
 
 function getPrefix(commandLine: string, cursorPosition: number): string {
 	// Return an empty string if the command line is empty after trimming
@@ -231,139 +196,6 @@ export function asArray<T>(x: T | T[]): T[] {
 	return Array.isArray(x) ? x : [x];
 }
 
-export type SpecArg = Fig.Arg | Fig.Suggestion | Fig.Option | string;
-
-export async function collectCompletionItemResult(
-	command: Command,
-	parsedArguments: ArgumentParserResult,
-	prefix: string,
-	terminalContext: { commandLine: string; cursorPosition: number },
-	shellIntegrationCwd: vscode.Uri | undefined,
-	items: vscode.TerminalCompletionItem[]
-): Promise<{ filesRequested: boolean; foldersRequested: boolean } | undefined> {
-	let filesRequested = false;
-	let foldersRequested = false;
-
-	const addSuggestions = async (specArgs: SpecArg[] | Record<string, SpecArg> | undefined, kind: vscode.TerminalCompletionItemKind, parsedArguments?: ArgumentParserResult) => {
-		if (kind === vscode.TerminalCompletionItemKind.Argument && parsedArguments?.currentArg?.generators) {
-			const generators = parsedArguments.currentArg.generators;
-			for (const generator of generators) {
-				if (generator.template) {
-					const templates = Array.isArray(generator.template) ? generator.template : [generator.template];
-					for (const template of templates) {
-						if (template === 'filepaths') {
-							filesRequested = true;
-						} else if (template === 'folders') {
-							foldersRequested = true;
-						}
-					}
-				} else {
-					const initialFigState: FigState = {
-						buffer: terminalContext.commandLine,
-						cursorLocation: terminalContext.cursorPosition,
-						cwd: shellIntegrationCwd?.fsPath ?? null,
-						processUserIsIn: null,
-						sshContextString: null,
-						aliases: {},
-						environmentVariables: {},
-						shellContext: undefined,
-					};
-					const state: AutocompleteState = {
-						figState: initialFigState,
-						parserResult: parsedArguments,
-						generatorStates: [],
-						command,
-
-						visibleState: Visibility.HIDDEN_UNTIL_KEYPRESS,
-						lastInsertedSuggestion: null,
-						justInserted: false,
-
-						selectedIndex: 0,
-						suggestions: [],
-						hasChangedIndex: false,
-
-						historyModeEnabled: false,
-						fuzzySearchEnabled: false,
-						userFuzzySearchEnabled: false,
-					};
-					const s = createGeneratorState(state);
-					const generatorResults = s.triggerGenerators(parsedArguments);
-					for (const generatorResult of generatorResults) {
-						for (const item of (await generatorResult?.request) ?? []) {
-							if (!item.name) {
-								continue;
-							}
-							const suggestionLabels = getLabel(item);
-							if (!suggestionLabels) {
-								continue;
-							}
-							for (const label of suggestionLabels) {
-								items.push(createCompletionItem(
-									terminalContext.cursorPosition,
-									prefix,
-									{ label },
-									undefined,
-									typeof item === 'string' ? item : item.description,
-									kind
-								));
-							}
-						}
-					}
-				}
-			}
-		}
-		if (!specArgs) {
-			return { filesRequested, foldersRequested };
-		}
-
-		if (Array.isArray(specArgs)) {
-			for (const item of specArgs) {
-				const suggestionLabels = getLabel(item);
-				if (!suggestionLabels) {
-					continue;
-				}
-				for (const label of suggestionLabels) {
-					items.push(
-						createCompletionItem(
-							terminalContext.cursorPosition,
-							prefix,
-							{ label },
-							undefined,
-							typeof item === 'string' ? item : item.description,
-							kind
-						)
-					);
-				}
-			}
-		} else {
-			for (const [label, item] of Object.entries(specArgs)) {
-				items.push(
-					createCompletionItem(
-						terminalContext.cursorPosition,
-						prefix,
-						{ label },
-						undefined,
-						typeof item === 'string' ? item : item.description,
-						kind
-					)
-				);
-			}
-		}
-	};
-
-	if (parsedArguments.suggestionFlags & SuggestionFlag.Args) {
-		await addSuggestions(parsedArguments.currentArg?.suggestions, vscode.TerminalCompletionItemKind.Argument, parsedArguments);
-	}
-	if (parsedArguments.suggestionFlags & SuggestionFlag.Subcommands) {
-		await addSuggestions(parsedArguments.completionObj.subcommands, vscode.TerminalCompletionItemKind.Method);
-	}
-	if (parsedArguments.suggestionFlags & SuggestionFlag.Options) {
-		await addSuggestions(parsedArguments.completionObj.options, vscode.TerminalCompletionItemKind.Flag);
-	}
-
-	return { filesRequested, foldersRequested };
-}
-
 export async function getCompletionItemsFromSpecs(
 	specs: Fig.Spec[],
 	terminalContext: { commandLine: string; cursorPosition: number },
@@ -390,7 +222,7 @@ export async function getCompletionItemsFromSpecs(
 	}
 
 	for (const spec of specs) {
-		const specLabels = getLabel(spec);
+		const specLabels = getFigSuggestionLabel(spec);
 
 		if (!specLabels) {
 			continue;
@@ -407,7 +239,7 @@ export async function getCompletionItemsFromSpecs(
 			// push it to the completion items
 			if (tokenType === TokenType.Command) {
 				if (availableCommand.kind !== vscode.TerminalCompletionItemKind.Alias) {
-					items.push(createCompletionItem(terminalContext.cursorPosition, prefix, { label: specLabel }, getDescription(spec), availableCommand.detail));
+					items.push(createCompletionItem(terminalContext.cursorPosition, prefix, { label: specLabel }, getFixSuggestionDescription(spec), availableCommand.detail));
 				}
 				continue;
 			}
@@ -424,20 +256,13 @@ export async function getCompletionItemsFromSpecs(
 				continue;
 			}
 
-			const command = getCommand(terminalContext.commandLine, {}, terminalContext.cursorPosition);
-			if (!command || !shellIntegrationCwd) {
-				continue;
-			}
-			const parsedArguments: ArgumentParserResult = await parseArguments(
-				command,
-				{ environmentVariables: env, currentWorkingDirectory: shellIntegrationCwd.fsPath, sshPrefix: '', currentProcess: name, /* TODO: pass in aliases */ },
-				spec,
-			);
-
-			const completionItemResult = await collectCompletionItemResult(command, parsedArguments, prefix, terminalContext, shellIntegrationCwd, items);
+			const completionItemResult = await getFigSuggestions(spec, terminalContext, prefix, shellIntegrationCwd, env, name, token);
 			if (completionItemResult) {
 				filesRequested ||= completionItemResult.filesRequested;
 				foldersRequested ||= completionItemResult.foldersRequested;
+				if (completionItemResult.items) {
+					items.push(...completionItemResult.items);
+				}
 			}
 		}
 	}
